requirements.txt
Django
django-model-admin-autocomplete